modif ajout livre

+++ backend/controllers/books.js
@@ -1,14 +1,33 @@
 // controllers/booksController.js
 const Book = require("../models/Book");
+const fs = require("fs");
 
-// Créer un nouveau livre
+const sharp = require("sharp");
+const path = require("path");
+
 exports.createBook = (req, res, next) => {
-	delete req.body._id;
-	const book = new Book({ ...req.body });
+	try {
+		const bookObject = JSON.parse(req.body.book);
+		delete bookObject._id;
+		delete bookObject._userId;
 
-	book.save()
-		.then(() => res.status(201).json({ message: "Livre enregistré !" }))
-		.catch((error) => res.status(400).json({ error }));
+		// Extraire juste le nom de fichier à partir du chemin compressé
+		const compressedFilename = req.file.compressedPath.split("/").pop();
+
+		const book = new Book({
+			...bookObject,
+			userId: req.auth.userId,
+			imageUrl: `${req.protocol}://${req.get(
+				"host"
+			)}/images/compressed/${compressedFilename}`,
+		});
+
+		book.save()
+			.then(() => res.status(201).json({ message: "Livre enregistré !" }))
+			.catch((error) => res.status(400).json({ error }));
+	} catch (error) {
+		res.status(400).json({ error });
+	}
 };
 
 // Modifier un livre
@@ -20,9 +39,26 @@ exports.updateBook = (req, res, next) => {
 
 // Supprimer un livre
 exports.deleteBook = (req, res, next) => {
-	Book.deleteOne({ _id: req.params.id })
-		.then(() => res.status(200).json({ message: "Livre supprimé !" }))
-		.catch((error) => res.status(400).json({ error }));
+	Book.findOne({ _id: req.params.id })
+		.then((thing) => {
+			if (thing.userId != req.auth.userId) {
+				res.status(401).json({ message: "Not authorized" });
+			} else {
+				const filename = thing.imageUrl.split("/images/")[1];
+				fs.unlink(`images/${filename}`, () => {
+					Book.deleteOne({ _id: req.params.id })
+						.then(() => {
+							res.status(200).json({
+								message: "Objet supprimé !",
+							});
+						})
+						.catch((error) => res.status(401).json({ error }));
+				});
+			}
+		})
+		.catch((error) => {
+			res.status(500).json({ error });
+		});
 };
 
 // Récupérer tous les livres
